# Conflicts:
#	server/src/main/java/org/apache/iotdb/db/qp/executor/PlanExecutor.java
#	server/src/main/java/org/apache/iotdb/db/qp/physical/crud/InsertPlan.java

+++ cluster/src/main/java/org/apache/iotdb/cluster/server/DataClusterServer.java
@@ -603,7 +603,7 @@ public class DataClusterServer extends RaftServer implements TSDataService.Async
   }
 
   @Override
-  public void onSnapshotApplied(Node header, List<Long> slots,
+  public void onSnapshotApplied(Node header, List<Integer> slots,
       AsyncMethodCallback<Boolean> resultHandler) {
     DataGroupMember dataMember = getDataMember(header, resultHandler, "Snapshot applied");
     dataMember.onSnapshotApplied(header, slots, resultHandler);
